+++ SpringCloudDev/springcloud-service-api/springcloud-system-api/src/main/java/org/springclouddev/system/vo/MenuVO.java
@@ -25,12 +25,12 @@ public class MenuVO extends Menu implements INode {
 	/**
 	 * 主键ID
 	 */
-	private Integer id;
+	private Long id;
 
 	/**
 	 * 父节点ID
 	 */
-	private Integer parentId;
+	private Long parentId;
 
 	/**
 	 * 子孙节点
